refactor(facade): split root lib.rs into exports/compat/prelude layers
Extract 170+ `pub use` re-exports from the monolithic src/lib.rs
(1,475 lines) into organized sub-modules:

- src/exports.rs + src/exports/*.rs: domain-organized re-exports
  (macros, settings, core_types, database, auth, rest, middleware,
  routing, views, forms, di, misc)
- src/compat.rs + src/compat/websockets.rs: WebSocketRouter WASM shim
- src/prelude.rs: extracted prelude (WASM stub + native server)

All `#[doc(hidden)]` macro-support modules (reinhardt_apps, reinhardt_di,
reinhardt_core, etc.) and `pub mod` wrapper declarations stay in lib.rs
to preserve macro-emitted path stability (D1/D2 constraints verified
via crate_paths.rs in 6 macro crates).

The `reinhardt_apps` and `urls` WASM shims also stay inline in lib.rs
because they create module-level namespace entries that macro path
resolution depends on (D4 constraint).

Public API surface is preserved via `pub use exports::*;` at the crate
root. Explicit re-exports for `Response` and `Header` in exports.rs
disambiguate names that appear in both core and OpenAPI glob re-exports.

Refs #4362
