+++ src/compat.rs
@@ -1,0 +1,3 @@
+//! WASM compatibility shims for native-only types.
+
+pub mod websockets;
