+++ src/exports/misc.rs
@@ -1,0 +1,16 @@
+//! Miscellaneous re-exports (tasks, test, storage, cache).
+
+#[cfg(feature = "tasks")]
+pub use reinhardt_tasks::{Scheduler, Task, TaskExecutor, TaskQueue};
+
+#[cfg(feature = "test")]
+pub use reinhardt_test::{APIClient, APIRequestFactory, APITestCase, TestResponse};
+
+#[cfg(feature = "storage")]
+pub use reinhardt_utils::storage::{InMemoryStorage, LocalStorage, Storage};
+
+#[cfg(feature = "cache")]
+pub use reinhardt_utils::cache::{Cache, CacheKeyBuilder, InMemoryCache};
+
+#[cfg(all(feature = "cache", feature = "redis-backend"))]
+pub use reinhardt_utils::cache::RedisCache;
